refactor: consolidate thread domain policy
